feat: setup firebase at gatsby-browser
