+++ package.json
@@ -15,7 +15,7 @@
     "serve": "gatsby serve"
   },
   "dependencies": {
-    "firebase": "^7.2.2",
+    "firebase": "^7.3.0",
     "gatsby": "^2.17.11",
     "gatsby-image": "^2.2.31",
     "gatsby-plugin-preact": "^3.1.20",
